+++ src/repr/adjacency_list/mod.rs
@@ -42,7 +42,7 @@
 //! ## Self-loop
 //!
 //! A self-loop is not allowed. The following pseudograph can not be
-//! represented. The self-loop is marked in red.
+//! represented. The self-loop is red:
 //!
 //! ![self-loop](https://raw.githubusercontent.com/bsdrks/graaf-images/main/out/adjacency_list_self_loop.svg?)
 //!
@@ -68,7 +68,7 @@
 //! ## Parallel arcs
 //!
 //! Parallel arcs are not allowed. The following multigraph can not be
-//! represented. The parallel arc is marked in red:
+//! represented. The parallel arc is red:
 //!
 //! ![parallel arcs](https://raw.githubusercontent.com/bsdrks/graaf-images/main/out/adjacency_list_parallel_arcs.svg?)
 //!